fix(webgl): fix NAPI buffer OOB read and integer overflow in WebGL paths
Add buffer bounds validation for srcOffset/length against JS ArrayBuffer
in texImage2D/texSubImage2D/compressedTexImage2D/compressedTexSubImage2D/
bufferData; use 64-bit arithmetic to prevent truncation/overflow; fix
CompressedTexImage2D PBO path arg index and unchecked napi status; fix
GenImageSource return code (false->GL_INVALID_ENUM); fix new[]/free
mismatch in external arraybuffer finalize; cap JS array alloc size;
change AllocBuffer param to size_t to prevent 4GB truncation bypass of
1MB cap; harden CheckCompressedTexData with overflow-checked helpers.

Found by deep security scan of interfaces/kits/napi/graphic/webgl.



🤖 AI[100%] 🔧 Human Fixed[0%] 🧑 Human[0%] 👌 AI Adopted[100%]

+++ interfaces/kits/napi/graphic/webgl/include/webgl/webgl_arg.h
@@ -555,9 +555,9 @@ public:
         }
     }
 
-    uint8_t* AllocBuffer(uint32_t size)
+    uint8_t* AllocBuffer(size_t size)
     {
-        if (size > MAX_ALLOC_BUFFER_SIZE) {
+        if (size > static_cast<size_t>(MAX_ALLOC_BUFFER_SIZE)) {
             return nullptr;
         }
         if (data_ != nullptr) {
